requirements.txt
Django
django-tables2
django-crispy-forms
crispy-bootstrap5
requests
pytz
tablib
django-import-export
zabbix-utils
